Moved the metric enum to the model package.
