+++ src/main/java/model/Metric.java
@@ -1,8 +1,8 @@
-package main.java.controller;
+package main.java.model;
 
 /**
  * Metric enumerator, for convinience. The current existing metrics are LOC
- * (Lines of Code), CYCLO (McCabe’s Cyclomatic Number), ATFD (Access To Foreign
+ * (Lines of Code), CYCLO (McCabeï¿½s Cyclomatic Number), ATFD (Access To Foreign
  * Data) and LAA (Locality of Attribute Accesses).
  *
  * @author Hugo Barroca
